+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.15]
+## Added
+- Emit `metadataParsed` and `metadata` events for generated EMSG/ID3 tags in the DateRangeEmitter.
+- Support for companion ads that have multiple `variations`.
+
 ## [1.2.14]
 ## Changed
 - Update `bitmovin-player` to version `8.45.1`